+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/interfaces/DeliveryStateNotifier.java
@@ -1,5 +1,5 @@
 package com.polytech.si5.al.dronedelivery.team.g.truck.interfaces;
 
 public interface DeliveryStateNotifier {
-    void updateDeliverySate(int droneId,int status);
+    void updateDeliverySate(long droneId,int status);
 }
