+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/interfaces/DroneStateNotifier.java
@@ -1,5 +1,5 @@
 package com.polytech.si5.al.dronedelivery.team.g.truck.interfaces;
 
 public interface DroneStateNotifier {
-    public void droneDown(int droneId);
+    public void droneDown(long droneId);
 }
